refactor(checker): split argument diagnostics helpers
